+++ app/src/main/java/com/dvinc/notepad/data/database/NotepadDatabase.kt
@@ -7,8 +7,8 @@ package com.dvinc.notepad.data.database
 
 import android.arch.persistence.room.Database
 import android.arch.persistence.room.RoomDatabase
-import com.dvinc.notepad.data.database.dao.Note
-import com.dvinc.notepad.data.database.entity.NoteDao
+import com.dvinc.notepad.data.database.entity.Note
+import com.dvinc.notepad.data.database.dao.NoteDao
 
 @Database(entities = arrayOf(Note::class), version = 1)
 abstract class NotepadDatabase : RoomDatabase() {
