+++ app/src/main/java/com/dvinc/notepad/data/database/entity/Note.kt
@@ -3,7 +3,7 @@
  * All rights reserved.
  */
 
-package com.dvinc.notepad.data.database.dao
+package com.dvinc.notepad.data.database.entity
 
 import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
@@ -11,7 +11,7 @@ import android.arch.persistence.room.PrimaryKey
 
 @Entity(tableName = "Notes")
 data class Note(
-        @PrimaryKey(autoGenerate = true) var id: Long = 0,
+        @PrimaryKey(autoGenerate = true) var id: Long,
         @ColumnInfo(name = "name") val name: String,
         @ColumnInfo(name = "content") val content: String,
         @ColumnInfo(name = "update_time") val updateTime: Long)
