+++ app/src/main/java/com/dvinc/notepad/ui/notepad/NotepadView.kt
@@ -5,11 +5,11 @@
 
 package com.dvinc.notepad.ui.notepad
 
-import com.dvinc.notepad.data.database.dao.Note
+import com.dvinc.notepad.data.database.entity.Note
 import com.dvinc.notepad.ui.base.MvpView
 
 interface NotepadView : MvpView {
-    fun showEmptyView();
+    fun showEmptyView()
     fun showError(message: String)
     fun showNotes(notes: List<Note>)
 }
